new files for the faq and form sections
