+++ src/components/ContactSection/ContactUsSection.tsx
@@ -1,48 +1,15 @@
 "use client";
 
-import React, { useState } from "react";
-import styles from "../../styles/ContactSection/ContactSection.module.css";
-import Link from "next/link";
+import React from "react";
+import Faq from "../FAQ/Faq";
+import Form from "../Form/Form";
 
-interface QAItem {
-  id: number;
-  question: string;
-  answer: string;
-}
 
-const qaData: QAItem[] = [
-  {
-    id: 1,
-    question: "How can I get started?",
-    answer:
-      "Getting started is easy! Sign up for an account, and you'll have access to our platform's features. No credit card required for the initial signup.",
-  },
-  {
-    id: 2,
-    question: "What is the pricing structure?",
-    answer:
-      "Our pricing structure is flexible. We offer both free and paid plans. You can choose the one that suits your needs and budget.",
-  },
-  {
-    id: 3,
-    question: "What kind of support do you provide?",
-    answer:
-      "We offer comprehensive customer support. You can reach out to our support team through various channels, including email, chat, and a knowledge base.",
-  },
-  {
-    id: 4,
-    question: "Can I cancel my subscription anytime?",
-    answer:
-      "Yes, you can cancel your subscription at any time without any hidden fees. We believe in providing a hassle-free experience for our users.",
-  },
-];
+
+
 
 const ContactUsSection: React.FC = () => {
-  const [visibleId, setVisibleId] = useState<number | null>(null);
 
-  const handleToggle = (id: number) => {
-    setVisibleId(visibleId === id ? null : id);
-  };
 
   return (
     <main className="w-full h-full max-w-[1350px] mx-auto">
@@ -58,243 +25,10 @@ const ContactUsSection: React.FC = () => {
       <div className="overflow-hidden bg-white py-8 sm:py-16 sm:ml-16">
         <div className="mx-auto max-w-7xl px-6 lg:px-8">
           <div className="mx-auto grid max-w-2xl grid-cols-1 gap-x-8 gap-y-16 sm:gap-y-20 lg:mx-0 lg:max-w-none lg:grid-cols-2">
-            <div className="lg:pr-8 lg:pt-4">
-              <div className="lg:max-w-lg">
-                <p className="flex justify-center mt-2 text-2xl font-bold tracking-tight sm:text-2xl">
-                  Frequently Asked Questions
-                </p>
-                <dl className="mt-10 max-w-xl space-y-8 text-base leading-7 text-gray-600 lg:max-w-none">
-                  <div className="max-w-3xl mx-auto mt-8 space-y-4 md:mt-16">
-                    {qaData.map(({ id, question, answer }) => (
-                      <div
-                        key={id}
-                        className="transition-all duration-200 bg-white shadow-lg cursor-pointer hover:bg-gray-50"
-                      >
-                        <button
-                          type="button"
-                          onClick={() => handleToggle(id)}
-                          className="flex items-center justify-between w-full px-4 py-5 sm:p-6"
-                        >
-                          <span className="flex text-lg font-semibold text-[#1976D2]">
-                            {question}
-                          </span>
-                          <svg
-                            className={`w-6 h-6 text-gray-400 ${visibleId === id ? styles.rotate : ""}`}
-                            xmlns="http://www.w3.org/2000/svg"
-                            fill="none"
-                            viewBox="0 0 24 24"
-                            stroke="currentColor"
-                          >
-                            <path
-                              strokeLinecap="round"
-                              strokeLinejoin="round"
-                              strokeWidth={2}
-                              d="M19 9l-7 7-7-7"
-                            />
-                          </svg>
-                        </button>
-                        <div
-                          className={`px-4 pb-5 sm:px-6 sm:pb-6 ${visibleId === id ? "block" : "hidden"}`}
-                        >
-                          <p>{answer}</p>
-                        </div>
-                      </div>
-                    ))}
-                  </div>
-                </dl>
-              </div>
-              <div className="mt-10 flex justify-end gap-x-6">
-                <a
-                  href="#"
-                  className="text-sm font-semibold leading-6 text-[#1976D2]"
-                >
-                  See All <span aria-hidden="true">→</span>
-                </a>
-              </div>
-              <div>
-                <p className="flex justify-center mt-2 mb-10 text-3xl font-bold tracking-tight sm:text-2xl">
-                  Our Links
-                </p>
-                <Link
-                  className="flex justify-center items-center mb-4"
-                  href={"#"}
-                >
-                  <svg
-                    xmlns="http://www.w3.org/2000/svg"
-                    fill="none"
-                    viewBox="0 0 24 24"
-                    strokeWidth={1.5}
-                    stroke="currentColor"
-                    className="size-6 text-[#1976D2]"
-                  >
-                    <path
-                      strokeLinecap="round"
-                      strokeLinejoin="round"
-                      d="M21.75 6.75v10.5a2.25 2.25 0 0 1-2.25 2.25h-15a2.25 2.25 0 0 1-2.25-2.25V6.75m19.5 0A2.25 2.25 0 0 0 19.5 4.5h-15a2.25 2.25 0 0 0-2.25 2.25m19.5 0v.243a2.25 2.25 0 0 1-1.07 1.916l-7.5 4.615a2.25 2.25 0 0 1-2.36 0L3.32 8.91a2.25 2.25 0 0 1-1.07-1.916V6.75"
-                    />
-                  </svg>
-                  <span className="text-sm text-[#1976D2] ml-3">
-                    theinernplacejobs@gmail.com
-                  </span>
-                </Link>
-                <Link className="flex justify-center items-center" href={"#"}>
-                  <svg
-                    xmlns="http://www.w3.org/2000/svg"
-                    fill="none"
-                    viewBox="0 0 24 24"
-                    strokeWidth={1.5}
-                    stroke="currentColor"
-                    className="size-6 text-[#1976D2]"
-                  >
-                    <path
-                      strokeLinecap="round"
-                      strokeLinejoin="round"
-                      d="M2.25 6.75c0 8.284 6.716 15 15 15h2.25a2.25 2.25 0 0 0 2.25-2.25v-1.372c0-.516-.351-.966-.852-1.091l-4.423-1.106c-.44-.11-.902.055-1.173.417l-.97 1.293c-.282.376-.769.542-1.21.38a12.035 12.035 0 0 1-7.143-7.143c-.162-.441.004-.928.38-1.21l1.293-.97c.363-.271.527-.734.417-1.173L6.963 3.102a1.125 1.125 0 0 0-1.091-.852H4.5A2.25 2.25 0 0 0 2.25 4.5v2.25Z"
-                    />
-                  </svg>
+            <Faq />
 
-                  <span className="text-sm text-[#1976D2] ml-3">
-                    09032884485
-                  </span>
-                </Link>
-                <ul className="flex justify-center mt-5 space-x-4">
-                  <li className="bg-[#1976D2] h-10 w-10 rounded-full flex items-center justify-center shrink-0">
-                    <a href="#">
-                      <svg
-                        xmlns="http://www.w3.org/2000/svg"
-                        width="20px"
-                        height="20px"
-                        fill="white"
-                        viewBox="0 0 24 24"
-                      >
-                        <path
-                          d="M6.812 13.937H9.33v9.312c0 .414.335.75.75.75l4.007.001a.75.75 0 0 0 .75-.75v-9.312h2.387a.75.75 0 0 0 .744-.657l.498-4a.75.75 0 0 0-.744-.843h-2.885c.113-2.471-.435-3.202 1.172-3.202 1.088-.13 2.804.421 2.804-.75V.909a.75.75 0 0 0-.648-.743A26.926 26.926 0 0 0 15.071 0c-7.01 0-5.567 7.772-5.74 8.437H6.812a.75.75 0 0 0-.75.75v4c0 .414.336.75.75.75zm.75-3.999h2.518a.75.75 0 0 0 .75-.75V6.037c0-2.883 1.545-4.536 4.24-4.536.878 0 1.686.043 2.242.087v2.149c-.402.205-3.976-.884-3.976 2.697v2.755c0 .414.336.75.75.75h2.786l-.312 2.5h-2.474a.75.75 0 0 0-.75.75V22.5h-2.505v-9.312a.75.75 0 0 0-.75-.75H7.562z"
-                          data-original="#000000"
-                        />
-                      </svg>
-                    </a>
-                  </li>
-                  <li className="bg-[#1976D2] h-10 w-10 rounded-full flex items-center justify-center shrink-0">
-                    <a href="#">
-                      <svg
-                        xmlns="http://www.w3.org/2000/svg"
-                        width="20px"
-                        height="20px"
-                        fill="white"
-                        viewBox="0 0 511 512"
-                      >
-                        <path
-                          d="M111.898 160.664H15.5c-8.285 0-15 6.719-15 15V497c0 8.285 6.715 15 15 15h96.398c8.286 0 15-6.715 15-15V175.664c0-8.281-6.714-15-15-15zM96.898 482H30.5V190.664h66.398zM63.703 0C28.852 0 .5 28.352.5 63.195c0 34.852 28.352 63.2 63.203 63.2 34.848 0 63.195-28.352 63.195-63.2C126.898 28.352 98.551 0 63.703 0zm0 96.395c-18.308 0-33.203-14.891-33.203-33.2C30.5 44.891 45.395 30 63.703 30c18.305 0 33.195 14.89 33.195 33.195 0 18.309-14.89 33.2-33.195 33.2zm289.207 62.148c-22.8 0-45.273 5.496-65.398 15.777-.684-7.652-7.11-13.656-14.942-13.656h-96.406c-8.281 0-15 6.719-15 15V497c0 8.285 6.719 15 15 15h96.406c8.285 0 15-6.715 15-15V320.266c0-22.735 18.5-41.23 41.235-41.23 22.734 0 41.226 18.495 41.226 41.23V497c0 8.285 6.719 15 15 15h96.403c8.285 0 15-6.715 15-15V302.066c0-79.14-64.383-143.523-143.524-143.523zM466.434 482h-66.399V320.266c0-39.278-31.953-71.23-71.226-71.23-39.282 0-71.239 31.952-71.239 71.23V482h-66.402V190.664h66.402v11.082c0 5.77 3.309 11.027 8.512 13.524a15.01 15.01 0 0 0 15.875-1.82c20.313-16.294 44.852-24.907 70.953-24.907 62.598 0 113.524 50.926 113.524 113.523zm0 0"
-                          data-original="#000000"
-                        />
-                      </svg>
-                    </a>
-                  </li>
-                  <li className="bg-[#1976D2] h-10 w-10 rounded-full flex items-center justify-center shrink-0">
-                    <a href="#">
-                      <svg
-                        xmlns="http://www.w3.org/2000/svg"
-                        width="20px"
-                        height="20px"
-                        fill="white"
-                        viewBox="0 0 24 24"
-                      >
-                        <path d="M12 9.3a2.7 2.7 0 1 0 0 5.4 2.7 2.7 0 0 0 0-5.4Zm0-1.8a4.5 4.5 0 1 1 0 9 4.5 4.5 0 0 1 0-9Zm5.85-.225a1.125 1.125 0 1 1-2.25 0 1.125 1.125 0 0 1 2.25 0ZM12 4.8c-2.227 0-2.59.006-3.626.052-.706.034-1.18.128-1.618.299a2.59 2.59 0 0 0-.972.633 2.601 2.601 0 0 0-.634.972c-.17.44-.265.913-.298 1.618C4.805 9.367 4.8 9.714 4.8 12c0 2.227.006 2.59.052 3.626.034.705.128 1.18.298 1.617.153.392.333.674.632.972.303.303.585.484.972.633.445.172.918.267 1.62.3.993.047 1.34.052 3.626.052 2.227 0 2.59-.006 3.626-.052.704-.034 1.178-.128 1.617-.298.39-.152.674-.333.972-.632.304-.303.485-.585.634-.972.171-.444.266-.918.299-1.62.047-.993.052-1.34.052-3.626 0-2.227-.006-2.59-.052-3.626-.034-.704-.128-1.18-.299-1.618a2.619 2.619 0 0 0-.633-.972 2.595 2.595 0 0 0-.972-.634c-.44-.17-.914-.265-1.618-.298-.993-.047-1.34-.052-3.626-.052ZM12 3c2.445 0 2.75.009 3.71.054.958.045 1.61.195 2.185.419A4.388 4.388 0 0 1 19.49 4.51c.457.45.812.994 1.038 1.595.222.573.373 1.227.418 2.185.042.96.054 1.265.054 3.71 0 2.445-.009 2.75-.054 3.71-.045.958-.196 1.61-.419 2.185a4.395 4.395 0 0 1-1.037 1.595 4.44 4.44 0 0 1-1.595 1.038c-.573.222-1.227.373-2.185.418-.96.042-1.265.054-3.71.054-2.445 0-2.75-.009-3.71-.054-.958-.045-1.61-.196-2.185-.419A4.402 4.402 0 0 1 4.51 19.49a4.414 4.414 0 0 1-1.037-1.595c-.224-.573-.374-1.227-.419-2.185C3.012 14.75 3 14.445 3 12c0-2.445.009-2.75.054-3.71s.195-1.61.419-2.185A4.392 4.392 0 0 1 4.51 4.51c.45-.458.994-.812 1.595-1.037.574-.224 1.226-.374 2.185-.419C9.25 3.012 9.555 3 12 3Z"></path>
-                      </svg>
-                    </a>
-                  </li>
-                  <li className="bg-[#1976D2] h-10 w-10 rounded-full flex items-center justify-center shrink-0">
-                    <a href="#">
-                      <svg
-                        xmlns="http://www.w3.org/2000/svg"
-                        x="0px"
-                        y="0px"
-                        width="20"
-                        height="20"
-                        fill="white"
-                        viewBox="0 0 50 50"
-                      >
-                        <path d="M 11 4 C 7.1456661 4 4 7.1456661 4 11 L 4 39 C 4 42.854334 7.1456661 46 11 46 L 39 46 C 42.854334 46 46 42.854334 46 39 L 46 11 C 46 7.1456661 42.854334 4 39 4 L 11 4 z M 11 6 L 39 6 C 41.773666 6 44 8.2263339 44 11 L 44 39 C 44 41.773666 41.773666 44 39 44 L 11 44 C 8.2263339 44 6 41.773666 6 39 L 6 11 C 6 8.2263339 8.2263339 6 11 6 z M 13.085938 13 L 22.308594 26.103516 L 13 37 L 15.5 37 L 23.4375 27.707031 L 29.976562 37 L 37.914062 37 L 27.789062 22.613281 L 36 13 L 33.5 13 L 26.660156 21.009766 L 21.023438 13 L 13.085938 13 z M 16.914062 15 L 19.978516 15 L 34.085938 35 L 31.021484 35 L 16.914062 15 z"></path>
-                      </svg>
-                    </a>
-                  </li>
-                </ul>
-              </div>
-            </div>
-            <div className="shadow-lg ml-5 w-5/5 sm:w-4/5 mt-10">
-              <p className="mt-2 text-xl text-[#1976D2] font-bold tracking-tight sm:text-xl ml-9 sm:mt-10">
-                Send us a message
-              </p>
-              <form className="text w-4/5 ml-9 sm:mt-20">
-                <div className="p-2 w-full">
-                  <div className="relative">
-                    <label
-                      htmlFor="name"
-                      className="leading-7 py-4 text-xl font-bold tracking-tight sm:text-xl"
-                    >
-                      Full Name
-                    </label>
-                    <input
-                      type="text"
-                      id="name"
-                      name="name"
-                      required
-                      className=" w-full bg-white rounded border border-gray-200 focus:border-blue-500 focus:bg-white focus:ring-2 focus:ring-blue-200 text-base outline-none text-gray-900 py-1 px-1 leading-8 transition-colors duration-200 ease-in-out "
-                    />
-                  </div>
-                </div>
-                <div className="p-2 w-full">
-                  <div className="relative">
-                    <label
-                      htmlFor="email"
-                      className="leading-7 py-4 text-xl font-bold tracking-tight sm:text-xl"
-                    >
-                      Your Email
-                    </label>
-                    <input
-                      type="email"
-                      id="email"
-                      name="email"
-                      required
-                      className=" w-full bg-white rounded border border-gray-200 focus:border-blue-500 focus:bg-white focus:ring-2 focus:ring-blue-200 text-base outline-none text-gray-900 py-1 px-1 leading-8 transition-colors duration-200 ease-in-out"
-                    />
-                  </div>
-                </div>
-                <div className="p-2 w-full">
-                  <div className="relative">
-                    <label
-                      htmlFor="name"
-                      className="leading-7 py-4 text-xl font-bold tracking-tight sm:text-xl"
-                    >
-                      Company Name
-                    </label>
-                    <input
-                      type="text"
-                      id="name"
-                      name="company name"
-                      required
-                      className=" w-full bg-white rounded border border-gray-200 focus:border-blue-500 focus:bg-white focus:ring-2 focus:ring-blue-200 text-base outline-none text-gray-900 py-1 px-1 leading-8 transition-colors duration-200 ease-in-out"
-                    />
-                  </div>
-                </div>
-                <div className="p-2 w-full">
-                  <div className="relative">
-                    <label
-                      htmlFor="message"
-                      className="leading-7 py-4 text-xl font-bold tracking-tight sm:text-xl"
-                    >
-                      Message
-                    </label>
-                    <textarea
-                      id="message"
-                      name="message"
-                      required
-                      className=" w-full bg-white rounded border border-gray-200 focus:border-blue-500 focus:bg-white focus:ring-2 focus:ring-blue-200 h-32 text-base outline-none text-gray-900 py-1 px-3 resize-none leading-6 transition-colors duration-200 ease-in-out"
-                    ></textarea>
-                  </div>
-                </div>
-              </form>
-            </div>
+            <Form />
+
           </div>
         </div>
       </div>
